+++ addOns/automation/CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - The ascan job 'Scan All Header' GUI label
 
+### Fixed
+- Register plans run by -autorun to prevent NPEs when editing them
+
 ## [0.13.0] - 2022-02-25
 ### Fixed
 - Issue when adding or removing add-ons via the UI (Issue 7075)